Linter cleanup + injectMoreContent added
