+++ bricks/aem-hero/aem-hero.html
@@ -2,8 +2,6 @@
   <style>
     /* Not using shadowroot here, so "local" CSS rules must be 
      * prepended with aem-hero for isolation.
-     * TODO it might be better to read a .css file and inject
-    *  its content here, so that file-based CSS tools can be used.
      */
     aem-hero {
       display: block;
