+++ scripts/aem.js
@@ -545,39 +545,3 @@ export class Brick extends HTMLElement {
     this.observer.observe(this.shadowRoot, { childList: true, subtree: true });
   }
 }
-
-/**
- * Simpler brick using data-aem-selector attributes in the
- * HTML template to select the content to inject in it.
- */
-export class HtmlTemplateBrick extends HTMLElement {
-  connectedCallback() {
-    const id = this.tagName.toLowerCase();
-    const template = document.getElementById(id);
-    if (template) {
-      const content = template.content.cloneNode(true);
-
-      // Inject content from CSS selectors
-      content.querySelectorAll('*[data-aem-selector]').forEach(e => {
-        this.querySelectorAll(e.dataset.aemSelector)?.forEach(src => e.append(src));
-      });
-
-      // let derived classes get the content that they need,
-      // before we replace it with our template
-      if(this.prepareContent) {
-        this.prepareContent();
-      }
-
-      // Append content to shadow root or directly
-      if(template.getAttribute('shadowroot') === 'true') {
-        this.attachShadow({ mode: 'open' });
-        this.shadowRoot.append(content);
-        this.root = this.shadowRoot;
-      } else {
-        this.innerHTML = '';
-        this.append(content);
-        this.root = this;
-      }
-    }
-  }
-}
